Move UE1 games SHA1Sums to UE1GameDatabase.h, and add gameVersionString to GameLaunchInfo
